red: move coordinate data import to /scripts/data_loader
