+++ scripts/data_loader/load_coordinates.py
@@ -1,0 +1,8 @@
+import json
+
+
+def load_sensor_metadata(file_path):
+    with open(file_path, "r") as file:
+        data = json.load(file)
+    return data["sensors"]
+
